% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_adhoc.R
\name{core_adhoc}
\alias{core_adhoc}
\title{Computes ad-hoc core inflation using the subitem exclusion method}
\usage{
core_adhoc(core, data, change, weight, code, date)
}
\arguments{
\item{core}{A string specifying the inflation core exclusion method: ex0, ex1, ex2 or ex3.}

\item{data}{Dataframe object with inflation sub-items data.}

\item{change}{A string with the column name of the percentage change of the inflation sub-items.}

\item{weight}{A string with the column name of the weight of the inflation sub-items.}

\item{code}{A string with the column name of the code of the inflation sub-items as provided in \code{\link[=get_ipca]{get_ipca()}}.}

\item{date}{A string with the name of the date column.}
}
\value{
Calculated core inflation tibble.
}
\description{
Computes ad-hoc core inflation using the subitem exclusion method
}
\examples{
\dontrun{
library(magrittr)
ipca_ex0 <- get_ipca(table = 7060) \%>\%
group_desc()

dplyr::tibble(
"date" = ipca_ex0[["IPCA MoM"]][["date"]],
"mom" = ipca_ex0[["IPCA MoM"]][["value"]],
"weight" = ipca_ex0[["IPCA Weight"]][["value"]],
"codigo" = ipca_ex0[["IPCA MoM"]][["code"]]
) \%>\% core_adhoc("ex0", ., "mom", "weight", "codigo", "date")
}
}
