% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difusao.R
\name{difusao}
\alias{difusao}
\title{Indice de difusao do IPCA}
\usage{
difusao(data)
}
\arguments{
\item{data}{Tibble devolvido por \code{\link[=get_ipca]{get_ipca()}}.}
}
\value{
Tibble com as colunas \code{date} e \code{variacao}, esta ultima em pontos
percentuais de 0 a 100.
}
\description{
Percentual de subitens do IPCA com variacao de precos positiva no mes
(Subsecao 2.6 da Nota Tecnica 57 do BCB). Os pesos nao entram no calculo.
}
\details{
Apenas os subitens efetivamente pesquisados no mes sao considerados. Nas
bases do IBGE, subitens nao pesquisados aparecem com variacao registrada
como \code{'...'}, que \code{\link[=get_ipca]{get_ipca()}} converte para \code{NA} -- e nao para zero, o que
os contaria erroneamente como precos estaveis no denominador.
}
\examples{
\dontrun{
ipca <- get_ipca(inicio = "2020-01")
difusao(ipca)
}
}
\seealso{
\code{\link[=core_p55]{core_p55()}}
}
