% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipca_classes_2020}
\alias{ipca_classes_2020}
\title{IPCA/IBGE classifications (December/2019 Quarterly Inflation Report from BCB).}
\format{
A list with 8 numeric vectors.
\describe{
\item{monitorados}{tracked items}
\item{servicos}{service sector items}
\item{servicos_subjacente}{items from the underlying service sector}
\item{servicos_ex_subjacente}{ex-underlying service sector items}
\item{ex0}{IPCA-EX0 exclusion core items}
\item{ex1}{IPCA-EX1 exclusion core items}
\item{ex2}{IPCA-EX2 exclusion core items}
\item{ex3}{IPCA-EX3 exclusion core items}
\item{smoothed}{IPCA-MS items to smooth}
...
}
}
\source{
\url{https://www.bcb.gov.br/content/ri/relatorioinflacao/201912/ri201912p.pdf}
}
\usage{
ipca_classes_2020
}
\description{
A dataset containing sub-item codes and classifications for core
inflation exclusion methods.
}
\keyword{datasets}
