% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipca_ms_itens}
\alias{ipca_ms_itens}
\title{Itens com variacao suavizada no nucleo MS}
\format{
Tibble com 43 linhas e 2 colunas:
\describe{
\item{pof}{POF de referencia da estrutura do IPCA}
\item{codigo}{codigo estrutural do item suavizado}
}
}
\source{
Tabela 5 da Nota Tecnica 57 do BCB (dezembro/2025).
}
\usage{
ipca_ms_itens
}
\description{
Codigos dos itens cuja variacao mensal e substituida pela variacao media
dos ultimos 12 meses no calculo do nucleo de medias aparadas com suavizacao
(Tabela 5 da NT 57).
}
\details{
Sao nove itens nas tres estruturas mais recentes. Nas duas primeiras sao
oito, porque o item \verb{8104.Cursos diversos} ainda nao existia. Entre
janeiro de 1991 e julho de 1999, os itens \verb{7301.Educacao} e
\verb{5201.Comunicacoes} substituem \verb{8101.Cursos regulares} e \verb{9101.Comunicacao}.
}
\keyword{datasets}
