% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipca_proxies}
\alias{ipca_proxies}
\title{Proxies de compatibilizacao entre estruturas do IPCA}
\format{
Tibble com 50 linhas e 6 colunas:
\describe{
\item{nucleo}{nucleo ao qual a proxy se aplica: \code{"DP"} ou \code{"MS"}}
\item{pof_destino}{estrutura de destino, cujos meses iniciais usam a proxy}
\item{item}{codigo do item na estrutura de destino}
\item{rotulo}{nome do item na estrutura de destino}
\item{sinal}{\code{+1} se o componente soma na proxy, \code{-1} se subtrai}
\item{componente}{codigo do componente na estrutura anterior}
}
}
\source{
Tabelas 2 a 4 e 6 a 8 da Nota Tecnica 57 do BCB (dezembro/2025).
}
\usage{
ipca_proxies
}
\description{
Os nucleos MS e DP dependem do historico recente de variacoes de precos --
11 e 48 meses, respectivamente. Na transicao entre estruturas do IPCA, itens
sem correspondencia direta na estrutura anterior precisam de uma serie proxy,
definida como media ponderada de componentes da estrutura antiga com os
sinais indicados (Tabelas 2 a 4 para o DP e 6 a 8 para o MS).
}
\details{
As proxies sao usadas exclusivamente nos meses iniciais de vigencia de cada
estrutura: os primeiros 48 meses no DP e os primeiros 11 no MS. A transicao
de julho/2006-dezembro/2011 para janeiro/2012-dezembro/2019 nao exigiu
compatibilizacao e por isso nao aparece aqui.
}
\keyword{datasets}
