% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleos-package.R
\docType{package}
\name{nucleos-package}
\alias{nucleos}
\alias{nucleos-package}
\title{nucleos: Central Bank of Brazil Inflation Cores}
\description{
Calculates the inflation cores and other analytical series derived from the IPCA, following the consolidated methodology published by the Central Bank of Brazil in Technical Note 57 (December 2025). Covers the 22 series published in the SGS, from January 1991 onwards, spanning the five IPCA structures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/analisemacropro/nucleos}
  \item Report bugs at \url{https://github.com/analisemacropro/nucleos/issues}
}

}
\author{
\strong{Maintainer}: Fernando da Silva \email{fernando@fortietwo.com} (\href{https://orcid.org/0000-0002-7491-6930}{ORCID})

Authors:
\itemize{
  \item Fernando da Silva \email{fernando@fortietwo.com} (\href{https://orcid.org/0000-0002-7491-6930}{ORCID})
}

}
\keyword{internal}
