% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{series_sgs}
\alias{series_sgs}
\title{Codigos das series analiticas do IPCA no SGS}
\format{
Tibble com 22 linhas e 5 colunas:
\describe{
\item{serie}{nome da serie}
\item{sgs}{codigo da serie nova no SGS}
\item{sgs_anterior}{codigo da serie anterior; \code{NA} quando nao existe}
\item{funcao}{funcao do pacote que reproduz a serie}
\item{inicio}{primeiro mes da serie nova}
}
}
\source{
Tabelas 9 e 11 da Nota Tecnica 57 do BCB (dezembro/2025).
}
\usage{
series_sgs
}
\description{
Codigos das 22 series consolidadas pela Nota Tecnica 57 no Sistema
Gerenciador de Series Temporais do Banco Central, com a funcao do pacote que
reproduz cada uma. Usado pela suite de testes de conformidade.
}
\details{
As series novas mantiveram os codigos que ja estavam em vigor. As anteriores
continuam no SGS sob os codigos de \code{sgs_anterior}, mas deixaram de ser
atualizadas. Alimentacao no domicilio, MA, P55 e difusao nao mudaram alem da
extensao do historico e por isso nao ganharam codigo para a versao anterior.
}
\keyword{datasets}
