% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_adhoc.R
\name{core_adhoc}
\alias{core_adhoc}
\title{Title}
\usage{
core_adhoc(core, data, change, weight, date)
}
\arguments{
\item{core}{Inflation core}

\item{data}{Data frame}

\item{change}{Percent change}

\item{weight}{Inflation weight}

\item{date}{Date column}
}
\value{
Tibble
}
\description{
Title
}
