% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_dp.R
\name{core_dp}
\alias{core_dp}
\title{Computes core inflation using the double weight method}
\usage{
core_dp(data, change, weight, code, group, date)
}
\arguments{
\item{data}{Dataframe object with inflation index and items data.}

\item{change}{A string with the column name of the percentage change of the inflation index and items.}

\item{weight}{A string with the column name of the weight of the inflation index and items.}

\item{code}{A string with the column name of the code of the inflation index and items as provided in \code{\link[=get_ipca]{get_ipca()}}.}

\item{group}{A string with the column name of the group of the inflation index and items as provided in \code{\link[=group_desc]{group_desc()}}.}

\item{date}{A string with the name of the date column.}
}
\value{
Calculated core inflation tibble.
}
\description{
Computes core inflation using the double weight method
}
\examples{
\dontrun{
library(magrittr)

ipca_dp <- get_ipca(period = "all") \%>\%
group_desc() \%>\%
core_dp(., "pct_change", "weight", "code", "group", "date")
}
}
\author{
Fernando da Silva <\href{mailto:fernando@fortietwo.com}{fernando@fortietwo.com}>
}
