% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ma.R
\name{core_ma}
\alias{core_ma}
\title{Computes core inflation using the non-smoothed trimmed means method}
\usage{
core_ma(data, change, weight, date)
}
\arguments{
\item{data}{Dataframe object with inflation items data.}

\item{change}{A string with the column name of the percentage change of the inflation items.}

\item{weight}{A string with the column name of the weight of the inflation items.}

\item{date}{A string with the name of the date column.}
}
\value{
Calculated core inflation tibble.
}
\description{
Computes core inflation using the non-smoothed trimmed means method
}
\examples{
\dontrun{
library(magrittr)
library(dplyr)

ipca_ma <- get_ipca(period = "all") \%>\%
group_desc() \%>\%
filter(group == "Item") \%>\%
core_ma(., "pct_change", "weight", "date")
}
}
