% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_ms.R
\name{core_ms}
\alias{core_ms}
\title{Computes core inflation using the smoothed trimmed means method}
\usage{
core_ms(data, change, weight, code, date)
}
\arguments{
\item{data}{Dataframe object with inflation items data.}

\item{change}{A string with the column name of the percentage change of the inflation items.}

\item{weight}{A string with the column name of the weight of the inflation items.}

\item{code}{A string with the column name of the code of the inflation items as provided in \code{\link[=get_ipca]{get_ipca()}}.}

\item{date}{A string with the name of the date column.}
}
\value{
Calculated core inflation tibble.
}
\description{
Computes core inflation using the smoothed trimmed means method
}
\examples{
\dontrun{
library(magrittr)
library(dplyr)

ipca_ms <- get_ipca(period = "all") \%>\%
group_desc() \%>\%
dplyr::filter(group == "Item") \%>\%
core_ms(., "pct_change", "weight", "code", "date")
}
}
\author{
Fernando da Silva <\href{mailto:fernando@fortietwo.com}{fernando@fortietwo.com}>
}
