% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipca.R
\name{get_ipca}
\alias{get_ipca}
\title{Fetch IPCA table's from SIDRA/IBGE}
\usage{
get_ipca(table = c(1419, 7060), period = "all")
}
\arguments{
\item{table}{A table from IBGE's SIDRA API.}

\item{period}{A character vector describing the period of data.}
}
\value{
List with data frames.
}
\description{
Fetch IPCA table's from SIDRA/IBGE
}
\examples{
\dontrun{
my_df <- get_sidra(1419, "last")
}
}
