% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_desc.R
\name{group_desc}
\alias{group_desc}
\title{Create IPCA groups, subsgroups, etc. descriptions and codes}
\usage{
group_desc(data)
}
\arguments{
\item{data}{Output from get_ipca()}
}
\value{
List of data frames with new columns
}
\description{
Create IPCA groups, subsgroups, etc. descriptions and codes
}
\examples{
\dontrun{
library(magrittr)
my_df <- get_sidra(period = "last") \%>\%
group_desc()
}
}
